introduces AssociateLogic and refactors the transition logic
